Moved utils folder to models for naming conventions
