+++ lib/src/models/item.dart
