+++ lib/src/models/product.dart
